feat: add the share button to the context menu
