+++ apps/client/src/app/pages/map/components/search-bar/components/context-menu/ContextMenu.tsx
@@ -3,7 +3,8 @@ import Menu from './components/Menu';
 function ContextMenu({
   setIsOpen,
   coordinates,
-  removeMarker
+  removeMarker,
+  feature
 }: ContextMenuProps) {
   return (
     <div className='absolute top-20 right-8 landscape:sm:right-20 landscape:md:right-8'>
@@ -11,6 +12,7 @@ function ContextMenu({
         coordinates={coordinates}
         setIsOpen={setIsOpen}
         removeMarker={removeMarker}
+        feature={feature}
       />
     </div>
   );
