+++ apps/client/src/app/pages/map/components/search-bar/components/context-menu/components/share-location/services/ShareLinkService.ts
